Lots of sculpt stuff.  I think I have a good idea how to design a better smooth brush with consistent "pressure". It's been driving me insane in Blender.
Basically laplacian smooth converges to a flat plane, i.e. what
clay brushes do.  What we want is for the amount of smoothing
to be proportional to the brush size, not geometric complexity.

So, basically we want a clay brush with autosmoothing, with the
autosmoothing subtracting from the clay brush's strength proportionally
to how much the smoothing moved verts.

If I get this working it would be extremely easy to port it to Blender.